refactor: fix tests WIP
