+++ backend/tsconfig.json
@@ -40,6 +40,8 @@
       "@email/*": ["email/*"],
       "@telegram/*": ["telegram/*"],
       "@shared/*": ["../../shared/src/*"],
+      "@models/*": ["../../shared/src/core/models/*"],
+      "@interfaces/*": ["../../shared/src/core/interfaces/*"],
       "@test-utils/*": ["test-utils/*"],
       "@database/*": ["database/*"],
       "@mocks/*": ["__mocks__/*"],
