+++ shared/jest.config.js
@@ -5,6 +5,7 @@ module.exports = {
   moduleNameMapper: {
     '@models/(.*)': '<rootDir>/src/core/models/$1',
     '@interfaces/(.*)': '<rootDir>/src/core/interfaces/$1',
+    '@core/(.*)': '<rootDir>/src/core/$1',
   },
   transform: {
     // eslint-disable-next-line @typescript-eslint/naming-convention
