+++ shared/src/core/models/email/email-op.ts
@@ -7,7 +7,7 @@ import {
   BelongsTo,
 } from 'sequelize-typescript'
 import { Campaign } from '@models/campaign'
-import { MessageStatus } from 'core/constants'
+import { MessageStatus } from '../../constants'
 
 @Table({ tableName: 'email_ops', underscored: true, timestamps: true })
 export class EmailOp extends Model<EmailOp> {
