+++ shared/src/core/models/list.ts
@@ -5,8 +5,8 @@ import {
   Table,
   BelongsToMany,
 } from 'sequelize-typescript'
-import { User, UserList } from 'core/models/index'
-import { ChannelType } from 'core/constants'
+import { User, UserList } from '@models/index'
+import { ChannelType } from '../constants'
 
 @Table({ tableName: 'lists', underscored: true, timestamps: true })
 // <Model<Partial<List>> is required to resolve errors in the model.create()
