+++ shared/src/core/models/protected_message.ts
@@ -6,7 +6,7 @@ import {
   Table,
   BelongsTo,
 } from 'sequelize-typescript'
-import { Campaign } from 'core/models/campaign'
+import { Campaign } from '@models/campaign'
 
 @Table({ tableName: 'protected_messages', underscored: true, timestamps: true })
 export class ProtectedMessage extends Model<ProtectedMessage> {
