+++ shared/src/core/models/user/user-credential.ts
@@ -6,8 +6,8 @@ import {
   ForeignKey,
 } from 'sequelize-typescript'
 import { User } from './user'
-import { Credential } from 'core/models/credential'
-import { ChannelType } from 'core/constants'
+import { Credential } from '@models/credential'
+import { ChannelType } from '../../constants'
 
 @Table({ tableName: 'user_credentials', underscored: true, timestamps: true })
 export class UserCredential extends Model<UserCredential> {
